APMrover2: Issue #11583 move contents of APMrover2.cpp to Rover.cpp
